+++ src/lib.rs
@@ -1,116 +1,12 @@
 #![allow(dead_code)]
 
+mod facilities;
+mod task;
+
 use authku::Client;
 use std::ops::Deref;
+use task::Task;
 
-pub struct Facility {
-    library: i32,
-    floor: i32,
-    facility_type: i32,
-    init_id: i32,
-    final_id: i32,
-    session_list: [&'static str; 20],
-}
-
-const DISCUSSION_ROOM: Facility = Facility {
-    library: 3,
-    floor: 3,
-    facility_type: 21,
-    init_id: 129,
-    final_id: 134,
-    session_list: [
-        "08300930", "09301030", "10301130", "11301230", "12301330", "13301430", "14301530",
-        "15301630", "16301730", "17301830", "18301930", "19302030", "20302200", "00000000",
-        "00000000", "00000000", "00000000", "00000000", "00000000", "00000000",
-    ],
-};
-
-const FCILITIES: [Facility; 1] = [DISCUSSION_ROOM];
-
-fn get_facility_by_id(id: i32) -> Option<&'static Facility> {
-    for facility in FCILITIES.iter() {
-        if id >= facility.init_id && id <= facility.final_id {
-            return Some(facility);
-        }
-    }
-    None
-}
-/// Task struct for booking
-pub struct Task {
-    date: String,
-    time: String,
-    session: String,
-    library: String,
-    floor: String,
-    facility_type: String,
-    facility_id: String,
-}
-
-impl Task {
-
-    pub fn new(date: &str, time: &str, facility_id: &str) -> Self {
-        let facility_id: i32 = facility_id.parse().unwrap();
-        let facility = get_facility_by_id(facility_id)
-            .unwrap_or_else(|| panic!("Invalid facility id: {} \nNot support id", facility_id));
-    
-        // get the idex of time in session_list
-        let mut session = 0;
-        for (i, t) in facility.session_list.iter().enumerate() {
-            if t == &time {
-                session = i;
-                break;
-            }
-        }
-    
-        Task {
-            date: date.to_string(),
-            time: time.to_string(),
-            session: session.to_string(),
-            library: facility.library.to_string(),
-            floor: facility.floor.to_string(),
-            facility_type: facility.facility_type.to_string(),
-            facility_id: facility_id.to_string(),
-        }
-    }
-
-    pub fn default(
-        date: &str,
-        time: &str,
-        session: &str,
-        library: &str,
-        floor: &str,
-        facility_type: &str,
-        facility_id: &str,
-    ) -> Self {
-        Task {
-            date: date.to_string(),
-            time: time.to_string(),
-            session: session.to_string(),
-            library: library.to_string(),
-            floor: floor.to_string(),
-            facility_type: facility_type.to_string(),
-            facility_id: facility_id.to_string(),
-        }
-    }
-
-
-    pub fn make_book_url(&self) -> String {
-        let url = format!("https://booking.lib.hku.hk/Secure/NewBooking.aspx?library={}&ftype={}&facility={}&date={}&session={}",
-            self.library,
-            self.facility_type,
-            self.facility_id,
-            self.date.replace("-", ""),
-            self.session,
-        );
-        url
-    }
-}
-
-impl From<(&str, &str, &str)> for Task {
-    fn from((date, time, facility_id): (&str, &str, &str)) -> Self {
-        Task::new(date, time, facility_id)
-    }
-}
 
 /// a wrapper of authku::Client
 pub struct LibClient {
@@ -173,7 +69,7 @@ impl LibClient {
                 .unwrap()
                 .as_str();
 
-        let session_key = format!("ctl00$main$listSession${}", &task.session);
+        let session_key = format!("ctl00$main$listSession${}", task.get_session());
 
         let mut data: [(&str, &str); 22] = [
             ("__VIEWSTATE", viewstate),
@@ -187,12 +83,12 @@ impl LibClient {
             ("__EVENTTARGET", ""),
             ("__EVENTARGUMENT", ""),
             ("__LASTFOCUS", ""),
-            ("ctl00$main$ddlLibrary", &task.library),
-            ("ctl00$main$ddlFloor", &task.floor),
-            ("ctl00$main$ddlType", &task.facility_type),
-            ("ctl00$main$ddlFacility", &task.facility_id),
-            ("ctl00$main$ddlDate", &task.date),
-            (&session_key, &task.time),
+            ("ctl00$main$ddlLibrary", task.get_library()),
+            ("ctl00$main$ddlFloor", task.get_floor()),
+            ("ctl00$main$ddlType", task.get_facility_type()),
+            ("ctl00$main$ddlFacility", task.get_facility_id()),
+            ("ctl00$main$ddlDate", task.get_date()),
+            (&session_key, task.get_time()),
             ("ctl00$main$txtUserDescription", ""),
             ("ctl00$main$hBtnSubmit", ""),
             ("ctl00$main$hBtnEmail", ""),
